backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Govind K <gok@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
semantic-kernel = "^1.3.0"
beautifulsoup4 = "^4.12.3"
azure-search-documents = "^11.4.0"
azure-ai-documentintelligence = "^1.0.0b4"
azure-identity = "^1.19.0"
openai = "^1.52.0"
tiktoken = "^0.8.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pyyaml = "^6.0.1"

ruff = "^0.8.0"
black = "^24.0.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
azure-ai-inference = "^1.0.0b4"
azure-monitor-opentelemetry = "^1.6.0"
numpy = "^1.24.0"
azure-ai-projects = "^1.0.0b12"
azure-ai-agents = "^1.1.0b2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
